requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
coloredlogs==15.0.1
flatbuffers==23.5.26
humanfriendly==10.0
idna==3.4
imageio==2.31.5
lazy_loader==0.3
llvmlite==0.41.0
mpmath==1.3.0
networkx==3.1
numba==0.58.0
numpy==1.25.2
onnxruntime==1.16.0
opencv-python-headless==4.8.1.78
packaging==23.2
Pillow==10.0.1
platformdirs==3.11.0
pooch==1.7.0
protobuf==4.24.4
PyMatting==1.1.10
pyreadline3==3.4.1
rembg==2.0.50
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.3
sympy==1.12
tifffile==2023.9.26
tqdm==4.66.1
urllib3==2.0.6
